Added support for orienting the thumbnail according to the orientation information in the EXIF metadata.
To enable this, the `ThumbnailParameter` class was made mutable so that the `ImageFilter` list that it contains can be changed to add filters to perform the required transformations to orient the thumbnail properly.

The `FileImageSource`- and `InputStreamImageSource` classes have been modified to call methods to obtain the orientation information from the EXIF utility classes, and add the required transformation filters to the `ThumbnailParameter`.

Also added some unit tests to check the behavior of the aforementioned classes.